requirements.txt
mcp[cli]>=1.0.0
httpx>=0.26.0
starlette>=0.36.3
uvicorn>=0.27.1
python-dotenv>=1.0.0 